--- .travis.yml
@@ -1,30 +1,0 @@
-language: python
-
-python:
-  - "2.7"
-  - "3.6"
-
-addons:
-  apt:
-    packages:
-      - oracle-java8-installer
-
-before_install:
-  - sudo update-java-alternatives -s java-8-oracle
-  - export JAVA_HOME=/usr/lib/jvm/java-8-oracle/jre
-  - java -version
-
-install:
-  - pip install tox-travis
-  - pip install codecov
-  - mkdir /tmp/elasticsearch
-  - wget -O - https://artifacts.elastic.co/downloads/elasticsearch/elasticsearch-5.5.0.tar.gz | tar xz --directory=/tmp/elasticsearch --strip-components=1
-  - /tmp/elasticsearch/bin/elasticsearch -d -E script.inline=true -E path.repo=/tmp -E repositories.url.allowed_urls='http://*' -E node.attr.testattr=test
-
-script: tox
-
-after_success:
-  - codecov
-
-notifications:
-  email: false
