+++ README.md
@@ -1,0 +1,7 @@
+# Logger ES handler
+
+## about
+
+Its a fork project from [CMRESHandler](https://github.com/cmanaha/python-elasticsearch-logger) module.
+
+Thanks for it Carlos Manzanedo Rueda!
