--- requirements/requirements_py27.txt
@@ -1,3 +1,0 @@
-elasticsearch==5.4.0
-requests==2.18.1
-enum==0.4.6
