--- requirements/requirements_py36.txt
@@ -1,2 +1,0 @@
-elasticsearch==5.4.0
-requests==2.18.1
