--- setup.cfg
@@ -1,1 +1,0 @@
-[bdist_wheel]
