--- sonar-project.properties
@@ -1,15 +1,0 @@
-sonar.projectKey=cmr.python:python-elasticsearch-logger
-sonar.projectName=Python Elasticsearch Logger
-sonar.projectVersion=1.0.0b4
-sonar.verbose=DEBUG
-sonar.language=py
-sonar.sources=cmreslogging
-sonar.tests=tests
-sonar.python.coverage.reportPath=coverage.xml
-sonar.sourceEncoding=UTF-8
-
-
-
-
-
-
