--- tox.ini
@@ -1,31 +1,0 @@
-[tox]
-envlist = {py27,py36}
-
-[testenv]
-passenv =
-    TEST_ES_SERVER
-    TEST_ES_PORT
-basepython =
-    py27: python2.7
-    py36: python3.6
-deps =
-    check-manifest
-    docutils
-    {py27}: readme_renderer
-    pylint
-    flake8
-    pytest
-    coverage
-commands =
-    python setup.py check -m -r -s
-    flake8 --max-line-length=120 cmreslogging/
-    #   pylint ./cmreslogging -r n --files-output=y '--msg-template="\{path\}:\{line\}: [\{msg_id\}(\{symbol\}), \{obj\}] \{msg\}"'
-    coverage erase
-    coverage run -a --source=./cmreslogging --branch tests/test_cmreshandler.py
-    coverage run -a --source=./cmreslogging --branch tests/test_cmresserializer.py
-    coverage xml -i
-    coverage html
-
-[flake8]
-exclude = .tox,*.egg,build,data
-select = E,W,F
